Add support for tcmalloc.
Control whether to delete all memory after the last problem with an
environment variable (VHPOP_FREE_ALL_MEMORY) instead of through a
compile-time option. Freeing all memory after planning has completed can
be expensive, as it requires traversing a lot of linked data structures.
We can save time by letting the operating system handle the cleanup.
However, to make the heapchecker happy, which is enabled in the
regression test when linked with tcmalloc, we need to free all memory
explicitly.